Add more feature_sharding
