+++ bcos-scheduler/src/SchedulerImpl.h
@@ -50,11 +50,7 @@ public:
         bcos::protocol::BlockFactory::Ptr blockFactory, bcos::txpool::TxPoolInterface::Ptr txPool,
         bcos::protocol::TransactionSubmitResultFactory::Ptr transactionSubmitResultFactory,
         bcos::crypto::Hash::Ptr hashImpl, bool isAuthCheck, bool isWasm, int64_t schedulerTermId,
-        size_t keyPageSize)
-      : SchedulerImpl(executorManager, ledger, storage, executionMessageFactory, blockFactory,
-            txPool, transactionSubmitResultFactory, hashImpl, isAuthCheck, isWasm, false,
-            schedulerTermId, keyPageSize)
-    {}
+        size_t keyPageSize);
 
     SchedulerImpl(ExecutorManager::Ptr executorManager, bcos::ledger::LedgerInterface::Ptr ledger,
         bcos::storage::TransactionalStorageInterface::Ptr storage,
@@ -62,46 +58,7 @@ public:
         bcos::protocol::BlockFactory::Ptr blockFactory, bcos::txpool::TxPoolInterface::Ptr txPool,
         bcos::protocol::TransactionSubmitResultFactory::Ptr transactionSubmitResultFactory,
         bcos::crypto::Hash::Ptr hashImpl, bool isAuthCheck, bool isWasm, bool isSerialExecute,
-        int64_t schedulerTermId, size_t keyPageSize)
-      : m_executorManager(std::move(executorManager)),
-        m_ledger(std::move(ledger)),
-        m_storage(std::move(storage)),
-        m_executionMessageFactory(std::move(executionMessageFactory)),
-        m_blockExecutiveFactory(
-            std::make_shared<bcos::scheduler::BlockExecutiveFactory>(isSerialExecute, keyPageSize)),
-        m_blockFactory(std::move(blockFactory)),
-        m_txPool(txPool),
-        m_transactionSubmitResultFactory(std::move(transactionSubmitResultFactory)),
-        m_hashImpl(std::move(hashImpl)),
-        m_isAuthCheck(isAuthCheck),
-        m_isWasm(isWasm),
-        m_isSerialExecute(isSerialExecute),
-        m_schedulerTermId(schedulerTermId),
-        m_preExeWorker("preExeScheduler", 2),  // assume that preExe is no slower than exe speed/2
-        m_exeWorker("exeScheduler", 1)
-    {
-        start();
-
-        if (!m_ledgerConfig)
-        {
-            std::promise<bcos::ledger::LedgerConfig::Ptr> promise;
-            auto future = promise.get_future();
-            asyncGetLedgerConfig(
-                [&promise](Error::Ptr const& error, bcos::ledger::LedgerConfig::Ptr ledgerConfig) {
-                    if (error)
-                    {
-                        SCHEDULER_LOG(ERROR) << LOG_DESC("failed to get ledger config")
-                                             << LOG_KV("error", error->errorCode())
-                                             << LOG_KV("errorMessage", error->errorMessage());
-                        promise.set_exception(std::make_exception_ptr(*error));
-                        return;
-                    }
-                    promise.set_value(std::move(ledgerConfig));
-                });
-
-            m_ledgerConfig = future.get();
-        }
-    }
+        int64_t schedulerTermId, size_t keyPageSize);
 
     SchedulerImpl(const SchedulerImpl&) = delete;
     SchedulerImpl(SchedulerImpl&&) = delete;
